// apps/studio/src/services/plugin/types.ts
interface BaseRequest {
  id: string;
}

interface GetQueryTextRequest extends BaseRequest {
  name: "getQueryText";
  args: void;
}

interface GetQueryTextWithCursorRequest extends BaseRequest {
  name: "getQueryTextWithCursor";
  args: void;
}

interface GetQueryResultsRequest extends BaseRequest {
  name: "getQueryResults";
  args: void;
}

interface ExecuteQueryInTabRequest extends BaseRequest {
  name: "executeQueryInTab";
  args: void;
}

interface InsertTextRequest extends BaseRequest {
  name: "insertText";
  args: {
    text: string;
  };
}

interface InsertTextAtCursorRequest extends BaseRequest {
  name: "insertTextAtCursor";
  args: {
    text: string;
  };
}

interface SwitchDatabaseRequest extends BaseRequest {
  name: "switchDatabase";
  args: {
    databaseName: string;
  };
}

interface GetTabListRequest extends BaseRequest {
  name: "getTabList";
  args: void;
}

interface SwitchToTabRequest extends BaseRequest {
  name: "switchToTab";
  args: {
    tabId: number;
  };
}

// Extended PluginRequestData type that includes internal methods
export type ExtendedPluginRequestData = 
  | PluginRequestData 
  | GetQueryTextRequest 
  | GetQueryTextWithCursorRequest 
  | GetQueryResultsRequest
  | ExecuteQueryInTabRequest
  | InsertTextRequest
  | InsertTextAtCursorRequest
  | SwitchDatabaseRequest
  | GetTabListRequest
  | SwitchToTabRequest;

/**
 * The kind of the tab. There is only one kind currently:
 *
 * - `base-tab`: A plain tab with no special UI.
 * - `shell-tab`: Like a query tab. This tab has two main parts; A plugin's
 *   `<iframe>` (placed at the top), and a table component (placed at the
 *   bottom). The table can be collapsed completely.
 **/
export type TabType = "shell" | "base";

export type PluginView = {
  /** The id of the view. */
  id: string;
  /** The name of the view that will be displayed in the UI */
  name: string;
  /** The type of the view. */
  type: `${TabType}-tab`;
  /** The html entry point of the view. For example, `index.html`. */
  entry: string;
};
// ...
export interface PluginMenuItem {
  /** The command or id of the menu item. This will be passed to the plugin. */
  command: string;

  /** User-facing label shown in the UI for this menu item. */
  name: string;

  /** The ID of a view defined in `capabilities.views`; the host opens a
   * new tab of that view. */
  view: string;

  /** Location or locations in the app where this menu item should appear,
   * expressed as a string or array of strings. */
  placement: PluginMenuItemPlacement | PluginMenuItemPlacement[];

  /** Optional group identifier for sorting and grouping items within a
   * placement.
   * @todo planned */
  group?: string;

  /** Optional numeric order for fine-grained sorting within a group, with
   * lower numbers shown first.
   * @todo planned */
  order?: number;
}

/** Used by earlier versions of AI Shell. */
type LegacyViews = {
  tabTypes?: {
    id: string;
    name: string;
    kind: TabType;
    /** Same as `entry` above. */
    entry: string;
  }[];
}

export type Manifest = ManifestV0 | ManifestV1;

export type ManifestV0 = {
  manifestVersion?: 0;
  id: string;
  name: string;
  author:
  | string
  | {
    name: string;
    url: string;
  };
  description: string;
  version: string;
  minAppVersion?: string;
  /** Requires a premium license (Ultimate edition) to use this plugin */
  requiresLicense?: boolean;
  /** Material UI icon name. https://fonts.google.com/icons?icon.set=Material+Icons */
  icon?: string;
  /** Provide all extension points here. */
  capabilities: {
    /** The list of views provided by the plugin. */
    views: LegacyViews;
    /** The list of menu items provided by the plugin. */
    menu: PluginMenuItem[];
  };
  /** The path to the plugin's root directory. This is helpful when you use a bundler to build the project to a `dist/` directory for example. */
  pluginEntryDir?: string;
  /** @todo not yet implemented. This is a list of settings that can be configured by config files. */
  settings?: {
    id: string;
    name: string;
    type: "string" | "number" | "boolean";
    description?: string;
    default: string | number | boolean;
  }[];
  /** @todo not yet implemented */
  permissions?: (
    | "run-custom-queries"
    | "create-entities"
    | "edit-entities"
  )[];
}

export type ManifestV1 = Omit<ManifestV0, "manifestVersion" | "capabilities"> & {
  manifestVersion: 1;
  capabilities: Omit<ManifestV0["capabilities"], "views"> & {
    views: PluginView[];
  }
};

export type PluginRegistryEntry = Pick<
  Manifest,
  "id" | "name" | "author" | "description"
> & {
  repo: string;
};

export interface Release {
  manifest: Manifest;
  sourceArchiveUrl: string;
}

export interface PluginRepository {
  latestRelease: Release;
  readme: string;
}

export type OnViewRequestListener = (params: OnViewRequestListenerParams) => void | Promise<void>;

export type OnViewRequestListenerParams = {
  source: HTMLIFrameElement;
  request: ExtendedPluginRequestData;
  after: (callback: (response: PluginResponseData) => void) => void;
  modifyResult: (callback: (result: PluginResponseData['result']) => PluginResponseData['result'] | Promise<PluginResponseData['result']>) => void;
}

export type PluginSettings = {
  [pluginId: string]: {
    autoUpdate: boolean;
    disabled?: boolean;
  }
}


export type WebPluginContext = {
  manifest: Manifest;
  store: PluginStoreService;
  utility: UtilityConnection;
  log: ReturnType<typeof rawLog.scope>;
  appVersion: string;
}

export type PluginContext = {
  manifest: Manifest;
  loadable: boolean;
}

export type WebPluginViewInstance = {
  iframe: HTMLIFrameElement;
  context: any; // PluginViewContext from @sqlmindstudio/plugin
}